Reorganised dropdown file structure
