+++ components/dropdown/dropdown-menu.ts
@@ -1,132 +1,14 @@
-import {Directive, Input, HostBinding, EventEmitter, Output, AfterContentInit, ContentChild, Renderer, ElementRef, HostListener, QueryList, ContentChildren, forwardRef} from '@angular/core';
+import {Directive, HostBinding, ContentChild, forwardRef, Renderer, ElementRef, AfterContentInit, ContentChildren, QueryList, Input, HostListener} from '@angular/core';
 import {SuiTransition, Transition} from '../transition/transition';
+import {DropdownService, DropdownAutoCloseType} from './dropdown.service';
 import {TransitionController} from '../transition/transition-controller';
-
-export type NewDropdownAutoCloseType = "itemClick" | "outsideClick" | "disabled";
-
-// Creates essentially a 'string' enum.
-export const NewDropdownAutoCloseType = {
-    ItemClick: "itemClick" as NewDropdownAutoCloseType,
-    OutsideClick: "outsideClick" as NewDropdownAutoCloseType,
-    Disabled: "disabled" as NewDropdownAutoCloseType
-}
-
-// Keyboard keycodes in use by the dropdown.
-export enum KeyCode {
-    Left = 37,
-    Up = 38,
-    Right = 39,
-    Down = 40,
-
-    Escape = 27,
-    Enter = 13,
-
-    Space = 32,
-    Backspace = 8
-};
-
-export class NewDropdownService {
-    // Open state of the dropdown
-    public isOpen:boolean;
-    // Emitter for when dropdown open state changes.
-    public isOpenChange:EventEmitter<boolean>;
-
-    public isDisabled:boolean;
-
-    // Sets the "autoclose" mode of the dropdown - i.e. what user action causes it to autoclose.
-    // Options are itemClick (close when choosing an item or clicking outside) [default], outsideClick (choose only when clicking outside) & disabled (never autoclose).
-    public autoCloseMode:NewDropdownAutoCloseType;
-
-    // Keep track of the containing dropdown so we can open it as necessary.
-    public parent:NewDropdownService;
-    // Also keep track of dropdowns nested in this one so we can close them as necessary.
-    public children:NewDropdownService[];
-    public get isNested() {
-        return !!this.parent;
-    }
-
-    constructor() {
-        this.isOpen = false;
-        this.isOpenChange = new EventEmitter<boolean>();
-
-        this.isDisabled = false;
-
-        this.autoCloseMode = NewDropdownAutoCloseType.ItemClick;
-
-        this.children = [];
-    }
-
-    public setOpenState(isOpen:boolean, reflectInParent:boolean = false) {
-        if (this.isOpen != isOpen && !this.isDisabled) {
-            // Only update the state if it has changed, and the dropdown isn't disabled.
-            this.isOpen = !!isOpen;
-            // We must delay the emitting to avoid the 'changed after checked' Angular errors.
-            this.delay(() => this.isOpenChange.emit(this.isOpen));
-
-            if (!this.isOpen) {
-                // Close the child dropdowns when this one closes.
-                this.children.forEach(c => c.setOpenState(this.isOpen));
-            }
-
-            if (this.parent && reflectInParent) {
-                // Open the parent dropdowns when this one opens.
-                this.parent.setOpenState(this.isOpen, true);
-            }
-        }
-        else if (this.isOpen != isOpen && this.isDisabled) {
-            // If the state has changed, but the dropdown is disabled, re-emit the original isOpen value.
-            this.delay(() => this.isOpenChange.emit(this.isOpen));
-        }
-    }
-
-    public setDisabledState(isDisabled:boolean) {
-        if (this.isDisabled != isDisabled) {
-            if (!!isDisabled) {
-                // Close the dropdown as it is now disabled
-                this.setOpenState(false);
-            }
-
-            this.isDisabled = !!isDisabled;
-        }
-    }
-
-    public toggleOpenState() {
-        this.setOpenState(!this.isOpen);
-    }
-
-    // Registers a dropdown service as a child of this service.
-    public registerChild(child:NewDropdownService) {
-        if (!this.isChildRegistered(child)) {
-            this.children.push(child);
-            child.parent = this;
-        }
-    }
-
-    // Recursive method to check if the provided dropdown is already registered as a child, or is a descendant of a child.
-    public isChildRegistered(child:NewDropdownService):boolean {
-        return this === child || !!this.children
-            .find(c => !!c.children
-                .find(c => c.isChildRegistered(child)));
-    }
-
-    // Wipes any nested data, so all services can be cleanly reattached.
-    public clearChildren() {
-        this.children.forEach(c => {
-            c.parent = null;
-        });
-        this.children = [];
-    }
-
-    // Method for delaying an event into the next tick, to avoid Angular "changed after checked" error.
-    private delay(callback:() => any) {
-        setTimeout(() => callback());
-    }
-}
+import {KeyCode} from '../util/util';
 
 @Directive({
+    // We must attach to every '.item' as Angular doesn't support > selectors.
     selector: '.item'
 })
-export class NewSuiDropdownMenuItem {
+export class SuiDropdownMenuItem {
     public get isDisabled() {
         // We must use nativeElement as Angular doesn't have a way of reading class information.
         const element = this._element.nativeElement as Element;
@@ -136,8 +18,8 @@ export class NewSuiDropdownMenuItem {
     @HostBinding('class.selected')
     public isSelected:boolean;
 
-    @ContentChild(forwardRef(() => NewSuiDropdownMenu))
-    public childDropdownMenu:NewSuiDropdownMenu;
+    @ContentChild(forwardRef(() => SuiDropdownMenu))
+    public childDropdownMenu:SuiDropdownMenu;
 
     public get hasChildDropdown() {
         return !!this.childDropdownMenu;
@@ -154,10 +36,10 @@ export class NewSuiDropdownMenuItem {
 }
 
 @Directive({
-    selector: '[newSuiDropdownMenu]'
+    selector: '[suiDropdownMenu]'
 })
-export class NewSuiDropdownMenu extends SuiTransition implements AfterContentInit {
-    private _service:NewDropdownService;
+export class SuiDropdownMenu extends SuiTransition implements AfterContentInit {
+    private _service:DropdownService;
     private _transitionController:TransitionController;
 
     // Allows the dropdown to be programmatically opened without being immediately closed by a mouse event.
@@ -167,7 +49,7 @@ export class NewSuiDropdownMenu extends SuiTransition implements AfterContentIni
         return this._service;
     }
 
-    public set service(value:NewDropdownService) {
+    public set service(value:DropdownService) {
         this._service = value;
 
         let previousIsOpen = this._service.isOpen;
@@ -187,8 +69,8 @@ export class NewSuiDropdownMenu extends SuiTransition implements AfterContentIni
         });
     }
 
-    @ContentChildren(NewSuiDropdownMenuItem)
-    private _items:QueryList<NewSuiDropdownMenuItem>;
+    @ContentChildren(SuiDropdownMenuItem)
+    private _items:QueryList<SuiDropdownMenuItem>;
 
     // Get the list of items, ignoring those that are disabled.
     public get items() {
@@ -196,7 +78,7 @@ export class NewSuiDropdownMenu extends SuiTransition implements AfterContentIni
     }
 
     // Stack that keeps track of the currently selected item. Selected items lower in the stack are necessarily the parent of the item one higher.
-    public selectedItems:NewSuiDropdownMenuItem[];
+    public selectedItems:SuiDropdownMenuItem[];
 
     // Sets whether or not to automatically select the 1st item when the dropdown is opened.
     @Input()
@@ -218,7 +100,7 @@ export class NewSuiDropdownMenu extends SuiTransition implements AfterContentIni
     public onClick(e:MouseEvent) {
         e.stopPropagation();
 
-        if (this._service.autoCloseMode == NewDropdownAutoCloseType.ItemClick) {
+        if (this._service.autoCloseMode == DropdownAutoCloseType.ItemClick) {
             if (e.srcElement.classList.contains("item")) {
                 // Once an item is selected, we can close the entire dropdown.
                 this._service.setOpenState(false, true);
@@ -235,7 +117,7 @@ export class NewSuiDropdownMenu extends SuiTransition implements AfterContentIni
     @HostListener("document:click", ["$event"])
     public onDocumentClick(e:MouseEvent) {
         if (this._isOpenOnMousedown) {
-            if (this._service.autoCloseMode == NewDropdownAutoCloseType.ItemClick || NewDropdownAutoCloseType.OutsideClick) {
+            if (this._service.autoCloseMode == DropdownAutoCloseType.ItemClick || DropdownAutoCloseType.OutsideClick) {
                 // No need to reflect in parent since they are also bound to document.
                 this._service.setOpenState(false);
             }
@@ -250,7 +132,7 @@ export class NewSuiDropdownMenu extends SuiTransition implements AfterContentIni
             // Gets the top selected item from the stack.
             let [selected] = this.selectedItems.slice(-1);
             // Keeping track of the menu containing the currently selected element allows us to easily determine its siblings.
-            let selectedContainer:NewSuiDropdownMenu = this;
+            let selectedContainer:SuiDropdownMenu = this;
             if (this.selectedItems.length >= 2) {
                 const [selectedParent] = this.selectedItems.slice(-2);
                 selectedContainer = selectedParent.childDropdownMenu;
@@ -302,7 +184,7 @@ export class NewSuiDropdownMenu extends SuiTransition implements AfterContentIni
     }
 
     // Determines the item to next be selected, based on the keyboard input & the currently selected item.
-    public updateSelection(selectedItem:NewSuiDropdownMenuItem, keyCode:KeyCode) {
+    public updateSelection(selectedItem:SuiDropdownMenuItem, keyCode:KeyCode) {
         if (selectedItem) {
             // Remove the selected status on the previously selected item.
             selectedItem.isSelected = false;
@@ -311,7 +193,7 @@ export class NewSuiDropdownMenu extends SuiTransition implements AfterContentIni
         let selectedIndex = this.items
             .findIndex(i => i === selectedItem);
 
-        let newSelection:NewSuiDropdownMenuItem;
+        let newSelection:SuiDropdownMenuItem;
 
         switch (keyCode) {
             case KeyCode.Enter:
@@ -357,94 +239,5 @@ export class NewSuiDropdownMenu extends SuiTransition implements AfterContentIni
             this.selectedItems.push(this._items.first);
         }
         
-    }
-}
-
-@Directive({
-    selector: '[newSuiDropdown]'
-})
-export class NewSuiDropdown implements AfterContentInit {
-    public service:NewDropdownService;
-
-    @ContentChild(NewSuiDropdownMenu)
-    private _menu:NewSuiDropdownMenu;
-
-
-    @ContentChildren(NewSuiDropdown, { descendants: true })
-    private _children:QueryList<NewSuiDropdown>;
-
-    public get children() {
-        // @ContentChildren includes the current element by default.
-        return this._children.filter(c => c !== this);
-    }
-
-    @Output()
-    public get isOpenChange() {
-        return this.service.isOpenChange;
-    }
-
-    @HostBinding('class.active')
-    public get isActive() {
-        // This is to ensure nested dropdowns don't get made bold.
-        return this.service.isOpen && !this.service.isNested;
-    }
-
-    @Input()
-    public get isOpen() {
-        return this.service.isOpen;
-    }
-
-    public set isOpen(value:boolean) {
-        // If we are opening the dropdown, we want to always open its parents.
-        this.service.setOpenState(value, !!value);
-    }
-
-    @HostBinding('class.disabled')
-    @Input()
-    public get isDisabled() {
-        return this.service.isDisabled;
-    }
-
-    public set isDisabled(value:boolean) {
-        this.service.setDisabledState(value);
-    }
-
-    @Input()
-    public get autoClose() {
-        return this.service.autoCloseMode;
-    }
-
-    public set autoClose(value:NewDropdownAutoCloseType) {
-        this.service.autoCloseMode = value;
-    }
-
-    constructor() {
-        this.service = new NewDropdownService();
-    }
-
-    public ngAfterContentInit() {
-        if (!this._menu) {
-            throw new Error("You must set [suiDropdownMenu] on the menu element.");
-        }
-        this._menu.service = this.service;
-
-        this.childrenUpdated();
-        this._children.changes
-            .subscribe(() => this.childrenUpdated());
-    }
-
-    private childrenUpdated() {
-        // Reregister child dropdowns each time the menu content changes.
-        this.children
-            .map(c => c.service)
-            .forEach(s => this.service.registerChild(s))
-    }
-
-    @HostListener("click", ['$event'])
-    public onClick(e:MouseEvent) {
-        // Block the click event from being fired on parent dropdowns.
-        e.stopPropagation();
-
-        this.service.toggleOpenState();
     }
 }
